BE: switch include processing to use FileId not SourceRootId
Summary:
In D64329884 we instroduced `db.file_app_data(file_id)` to look up the `SourceRootId` for a file and call `db.app_data(root_id)` on it.

This diff just switches to using `file_app_data()` in the local  include file resolution, so it passes a `FileId` not a `SourceRootId` for resolution.

Reviewed By: robertoaloi

Differential Revision: D65478074

fbshipit-source-id: 23ba507bfb75f7bc4aed863aecd171117bb5212f