Use Variable for base of dynamic() type semantic highlighting
Summary:
We use bold to highlight occurrences of `dynamic/0` or `dynamic/1` in the IDE.

The fact that it is dynamic is set as a modifier, not the base token type.

In the first version we set the base token type to `HlTag::None`, but this is not a "use the default base" setting.

Instead highlight as a variable, as this comprises 99% of our current cases.

Reviewed By: ilya-klyuchnikov

Differential Revision: D65601430

fbshipit-source-id: 3c3a60af5453e5d290a0799b12ad6ace72097ed4