Tracking down "stuck" diagnostics
Summary:
# Problem

We have a problem where sometimes we end up with stale diagnostics that do not clear.

This can be because we do not publish the updated diagnostics.

The normal process is to accept new diagnostics into the diagnostics store, and if they are different from what is already there, add the `FileId` to a list of ones to be re-published.

At the end of the server main loop event handler, we take this list and iterate over it, publishing diagnostics for all of those files.

This publish process takes a server `Snapshot`, and uses it for line index information for the files that it published.  Getting this line index has a `?` on it, because it can be cancelled.  If it *is* cancelled, the function returns early, and any unpublished changed files will now not be published, and the change is already registered in the diagnostic store so they will not be re-published until. they change.

# This diff

Detect that getting the line index errored, and if so return the list of changed files to the diagnostics collection, so they can be published next time around the loop.

Note: the cancellation should only be triggered on a write to the underlying salsa database inputs, which only happens at a specific point in the main loop.  So this scenario should not acttually occur.  But we check for it, and log a warning if it occurs.

Reviewed By: jcpetruzza

Differential Revision: D65671572

fbshipit-source-id: e6f417045c8ebe6a73d9d7afe1024f5e1f635ee9