Include source and root dirs in generated project model includes
This roughly matches what rebar3 does (see the link in the comment):
use `include/`, all of the source directories and the project root as
include directories. This only applies to the generated project model,
i.e. `.elp.toml` exists and defines `apps` and `dirs` and not `file`.

This fixes some `-include(..)` resolution in RabbitMQ, which uses the
generated build configuration project model and occasionally uses the
`-include("src/header.hrl")` or `-include("include/header.hrl")` idioms.