Add testcases for parametrized types
Summary:
The range for a "type call" currently includes the arguments. This is problematic in Glean/CodeHub where, in case of nested calls, it's not possible to navigate to the inner one.

This diff introduces a test case reproducing the issue, which will be solved later in the stack.

Notice how v1 of the indexing is not needed any longer and could be removed, but that's outside of the scope of this stack.

Reviewed By: alanz

Differential Revision: D66234608

fbshipit-source-id: f5b5b51e486c8680817e2bbed2786107e3802258