Read string representation of Eqwalizer diagnostic codes
Summary:
We currently render an Eqwalizer diagnostic code as a prefix `eqwalizer: ` followed by the string representation of the code from eqwalizer.

We are able to convert string representations of other diagnostic codes (ELP native, erlc, ad-hoc), but not eqwalizer ones.

This diff makes it possible to convert a string back into a diagnostic code, which is important as it allows us to specify them codes for `elp lint`, since we know have quick fixes attached.

For example, we can now do

```
 elp lint --include-eqwalizer-diagnostics --diagnostic-filter "eqwalizer: non_exported_id" --apply-fix --in-place
```

Reviewed By: robertoaloi

Differential Revision: D66242393

fbshipit-source-id: fb5f26bc99d90296d93d128457a090c355f94834