1/n: Split out buck processing for old and new
Summary:
This stack is about more closely aligning the internal project model with the buck target graph.

This is because there is a mismatch, arising from
- the existing model is based on the rebar3 concept of `applications`, where there is one directory per application, and it includes tests in a test directory, and may have dependencies on other applications.
- tn contrast, buck is organised around `targets`, where a target is an individual item that can be built and cached.

The implications of this are
- there can be multiple targets all co-existing withing a single directory.  It is common to have the main application as one, and then separate test targets for end-to-end and unit tests, each with different files included from the directory by regex, and each type having different dependencies.
- Test targets are generated per test file. So instead of a single set of tests, there are multiple test targets.

This means we need to move away from considering the `SourceRoot` as the primary index into the containing `AppData`.

We are also not currently making full use of buck's ability to generate new code based on input files, as we cannot represent this in the project model.

# This Diff

This diff clears the way for the rest of the work, by duplicating the buck processing path into old and new, depending on the `BuckQueryConfig` flag.  This either selects `Original` for the current processing, or `Bxl` for the new stack, which will make use of a buck bxl query to provide the target information.

Once we are comfortable that the new mechanism is working reliably, we can remove the old, but need both for the transition, to be able to fall back if there are issues.

Reviewed By: jcpetruzza

Differential Revision: D65941130

fbshipit-source-id: 92ae4284eab6de4c67396155820077f749bfa2e6