6/n: Add ProjectAppData.applicable_files field
Summary:
Rather than applying to all files in a `SourceRoot` directory, a buck target applies to a specific (sub)set of files.

In preparation for handling this, add a field to `ProjectAppData` containing this set of files, to be populated when it originates from a buck target.

The field is not used yet.

Reviewed By: robertoaloi

Differential Revision: D65943159

fbshipit-source-id: 13950711b02db0da38005f695b4034cb0732b703