8/n: Remove duplicates in Vfs loader config
Summary:
Our vfs loader config is based on the assumption that each application has a unique directory, and we make a list of these directories to load.

But with the buck bxl project model, especially test targets, we can have more than one application in the same root directory.

To avoid redundant load directories for the vfs loader, use a set of directories rather than a vector.

Note: this has an effect on the loading feedback, since we now only have 2 entries in the loader, one for directories and one for files.  We remedy this toward the end of the stack.

Reviewed By: robertoaloi

Differential Revision: D65944205

fbshipit-source-id: 44d1fcb05af835d5c7161e14fbe6f47443e4f8e0