9/n: Find app root based on BuckTarget info only
Summary:
The previous buck project model tried to synthesise an application per directory.  To do this, it would attempt to identify the root directory, using various heuristics such as if there is a `src`, `tests` or `includes` directory.

For the new project model, we already receive absolute paths from the bxl query, and can choose the highest common path of all the included source files.

We have a fallback to the base directory of the BUCK file containing the target, but that should never be used.

Reviewed By: robertoaloi

Differential Revision: D65946488

fbshipit-source-id: 38e1338e5de5ea7fd5806818bf0e0f485236f499