13:n: Create a map of FileId to AppDataId
Summary:
This diff brings in the map from `FileId` to `AppDataId`, and from `AppDataId` to `AppData`.

Because we will be loading files that do not necessarily appear in the `applicable_files` field for a given application, and it must also continue to work for the projects loaded via other means, we preserve the db function

```
fn app_data(&self, id: SourceRootId) -> Option<Arc<AppData>>;
```

but rather than having is as a `salsa::input`, it is a cached function instead.

We also introduce

```
    #[salsa::input]
    fn app_data_by_id(&self, id: AppDataId) -> Option<Arc<AppData>>;
    #[salsa::input]
    fn app_data_id(&self, id: SourceRootId) -> AppDataId;
```

Which allow us to move from `SourceRootId` to `AppData` as before.

The new functionality is

```
    fn app_data_id_by_file(&self, id: FileId) -> Option<AppDataId>;
```

which consults a table constructed from the `applicable_files` field for each application of `FileId` to `AppDataId`.  If this lookup fails it falls back to `FileId` -> SourceRootId` -> `AppData`

The nett effect is we can retrieve the required `AppData` by `FileId` if that data is populated, or fall back to the prior mechanism otherwise.

Reviewed By: robertoaloi

Differential Revision: D65951794

fbshipit-source-id: 63377dd7241db613eb7b2686699c799ad01aaf24