16/n: process included_apps for include path
Summary:
Where a buck target depends on includes from another one, but making it an `application` dependency would cause an illegal loop in the target graph, the include dependency is specified in `included_applications` instead, and already passed through from `elp.bxl`.

This diff reads the info from the `elp.bxl` json output, and uses it when calculating the include path for the target.

Reviewed By: robertoaloi

Differential Revision: D66173262

fbshipit-source-id: 77749e750abe66dad76e0c5df5e4a1d1ff6b8de0