15/n: compute module index based on file_id AppData
Summary:
We now have the `FileId` -> `AppData` mapping as the source of truth for a given file when loading a buck project, and `FileId` -> `SourceRoot` -> `AppData` as the fallback.

Update populating the `module index` to reflect this, by iterating over all source roots (every file has a source root), and for each file get its `AppData` via the new mechanism.

Reviewed By: robertoaloi

Differential Revision: D66169192

fbshipit-source-id: 159a167c551b4145fc3239659cdc8837e10ddb64