14/n add AppData.is_test_target, for buck test target apps
Summary:
For the rebar-based project model a target has both tests and source files, and the tests are in a specific directory.

For the new buck-based one, there is one buck target per test module, and the source module may be anywhere.

The bxl query we use tags them though, so we add a field to the `ProjectAppData` to capture this, as a `Option<bool>`, where the value is only set for the buck bxl project loading.

This allows us to correctly populate the module index for test targets too.

Reviewed By: robertoaloi

Differential Revision: D65998089

fbshipit-source-id: 91614f2cb45df286b39aef2a7b0a6fed2e497817