11/n: include test targets in project model
Summary:
Now that we have the application root dir accurately calculated, enable test targets as applications.

This still leaves the problem of using

`FileId` -> `SourceRoot` -> `AppData`

to get the `AppData` for a file, needed for include file resolution, macros, and so on.

Later parts of the stack move toward a

`FileId` -> `AppData`

path instead (via an intermediate `AppDataId`).

Reviewed By: robertoaloi

Differential Revision: D65949027

fbshipit-source-id: 3c4faf7af12420605b557ac8bf33323543bcb1de