28/n: break out make_buck_target
Summary:
We now have each buck target classified as being `app`, `dep` or `prelude`.  We intend only generating target configuration for the prelude targets actually used in the project, which means doing an initial pass over the `app` and `dep` ones, noting which `prelude` dependencies they have, and then generating config for those.

The net result is we need to use the target creation logic in two places further up the stack, so this diff breaks it out into a separate function in preparation.

Reviewed By: robertoaloi

Differential Revision: D66446074

fbshipit-source-id: 8528daecb853d1bd38db9f6d6edae67f0d7bb111