34/n: Provide --buck-bxl CLI arg for elp server
Summary:
We already have a configuration setting in the extension to use `elp.bxl` for buck2 queries.

This diff adds a CLI argument for `elp server`, allowing it to be used as `elp --buck-bxl server`.

This simplifies development testing, as the `ELP:Command` setting can use this, to easily run an old or new version, even if the extension is not yet updated for the new configuration.

Reviewed By: robertoaloi

Differential Revision: D66448377

fbshipit-source-id: db2e2bcc970a3866ef55cfa9276d7c5bf7780e5b