33/n: BE: using chunking in loader config for better progress reporting
Summary:
The original loader was based on the assumption that each application is wholly contained within a given directory, and that no applications overlap.

When the applications are generated from buck2 bxl, we have a mapping of one buck target to one application, and the directories may overlap, in particular for test targets, where there is one per test suite file.

To provide an efficient loading experience this stack has consolidated the list of directories and files to load, removing duplicates, and generates a single load configuration for
- all directories containing .erl and possibly .hrl files
- all directories containing only .hrl files
- all files listed as buck2 applicable files

Our progress bar feedback is tied to reporting on the loader configurations processed.  So it now has three steps only, one of which (.hrl files only) is very quick.

This diff brings back a more meaningful progress reporting bar by arbitrarily splitting the directory loading instructions into 100 chunks, split as 80 for .erl and .hrl directories, 20 for .hrl only.

These choices are arbitrary and not significant, purely to provide a bit more feedback in the porgress bar.

Reviewed By: robertoaloi

Differential Revision: D66448085

fbshipit-source-id: d6d5eed3bae2bb1a87ce077655d78efff6796661