29/n: Add BuckTarget origin field
Summary:
The previous diff updated `elp.bxl` to return a field giving the origin of a target as `app`, `dep`, or `prelude`.

Add a matching field to `BuckTarget`, including `serde` infrastructure to read and write this field from a string to an enum.

We will use this field later in the stack.

Reviewed By: robertoaloi

Differential Revision: D66306551

fbshipit-source-id: e993230ff4688533e2e70fb559185ea9997b3afb