35/n: set app_data_id for all files in IDE mode
Summary:
The basic project configuration process is as follows
- we generate a `Project` structure from whichever project description mechanism we use, e.g. rebar,  buck or a test fixture.
- From the project description, we generate our internal data structures for processing the project while running.  One of these, introduced in this stack, is a mapping from `FileId` to `AppData` (via `AppDataId`).
- Because of the way we abstract the file system, at the time we generate these structures we do not have `FileId`s, as these are generated by the file loading process.  For buck2 we do however have the applicable files list, and this is the reason we have the `FileId` to `AppData` mapping. For the rest of the `Project` description mechanisms knowing the directory of the file is enough, as we can look up `FileId` -> `SourceRootId` -> `AppData`.

For the CLI case, we load the entire project into VFS before generating its structure.  This means we *can* resolve the applicable file path to `FileId` while loading.

For the server, we do this the other way around.

# This diff

This diff modifies the internal data generation to return a map of file path to `AppDataId`, for later resolution once the loading is complete.

We store this list in the server state, and update the salsa data structure as the files are loaded.

Reviewed By: robertoaloi

Differential Revision: D66448798

fbshipit-source-id: f4684c8bc35cf75ec668784ab59bb962b5aa977a