31/n: use abs_src_dirs for buck, final CLI accuracy
Summary:
This ensures we only load the listed test `.erl` files, not the whole directory, and so avoid loading ancillary non-erlang test files.

`all_source_dirs` is used to generate the file loader instructions, and for the buck2 project model we populate `abs_src_dirs` only with those directories that are not `test`.

We use the presence of `BuckTarget.applicable_files` to detect that we are processing a buck2-originating target, so existing loader behaviour for rebar3 and other targets is unchanged.

This diff completes the accuracy of buck2 project model loading for the CLI case.  We address the IDE case in later diffs.

Reviewed By: robertoaloi

Differential Revision: D66447223

fbshipit-source-id: 8f41c79e8c71b4392eb8eee7e13cc2d5d64c6976