27/n: Update test results for elp.bxl change
Summary:
The `elp.bxl` script now tags each target with an origin of `app`, `dep`, or `prelude`, and includes all the prelude targets too.

Update the tests to show the change.

We do no further processing of these in ELP at this point, the changes happen further up the stack.

Reviewed By: robertoaloi

Differential Revision: D66445845

fbshipit-source-id: eb81e225c48276b13629859aaa4a3e4cda91ecba