Check for at least OTP 27 when running elp CLI commands that need it
Summary:
The ELP erlang service needs at least OTP27 on its path to function properly.

The project ELP is being run against can have a different PATH with a different OTP version, but ELP needs OTP27

Add a check to the CLI invocations, and bail if the condition is not met.

Reviewed By: michalmuskala

Differential Revision: D66580167

fbshipit-source-id: bd58404544d78206de531fc55a666b9c66b93d56