Auto-completion for maps
Summary:
This adds auto-completion for Erlang maps. They are mapped to "structs" in LSP terminology.

Map auto-completion is enabled for both `Expr` and `Other` contexts:
* `Expr` is used in function bodies
* `Other` enables auto-completion in function arguments (e.g. for matching)

We consider as candidates for auto-completion all the type aliases that resolve to maps defined in the current module or included header files. In addition, we scan the type specifications present in the current module (both corresponding to existing functions and not - known as *unowned*) to look for type aliases that also resolve to maps.

Notice how we set the `position` for the type definition in the completions. This will be used later to fetch type documentation for types when hovering a given auto-completion item.

Reviewed By: alanz

Differential Revision: D66543827

fbshipit-source-id: 36ceeb4a1aa3e075ec91d7add519b5e7a13cda61