Add support for type docs
Summary:
OTP 27 allows the usage of the `-doc` attribute not just for functions, but also for types (and callbacks, but those are out of scope for this change).

Here we extend the doc infrastructure to support types.

When presenting the doc for a type, we combine the type definition itself (since it could contain valuable comments) and eventual docs.

This also allows us to provide better auto-completion for maps.

Reviewed By: alanz

Differential Revision: D66578553

fbshipit-source-id: 9db32db97f77849f2b747f5018e4641a45dd6f77