Reliably process 'textDocument/documentSymbol' on startup
Summary:
The server startup process has essentially 3 phases
- `Initialize` message handline, which deals with capability negotiation between the client and server, and is complete before the server main loop is entered.
- Project discovery, and setting our internal data structures according to the project model
- Loading source files, as determined by the project model

Up to now, we have waited until all three are done before starting to process messages from the client (except shutdown request), returning error for all requests.

We do this to prevent errors occurring while processing the messages.

BUT, the errors arise from the project model not being set, not from the project loading being incomplete.

This also introduces the problem that the client sends some messages as soon as it gets the `initialize` response, and does not re-send them later unless the file changes.  This is particularly apparent with `textDocument/documentSymbol`, where the outline does not get populated until the file changes.

# This diff

This diff deals with document symbols on startup by
- Not initially registering that we can handle the requests (if dynamic registration for it is supported on the client)
- Dynamically registering support for it once the project model is setup up, but before loading is complete.
- Instead of giving an error response to all messages while loading, process `textDocument/documentSymbol`, since it is local to the file only, and the project model is set up so calls do not panic.

Reviewed By: robertoaloi

Differential Revision: D66574700

fbshipit-source-id: 74fb765a5b3e9411147200e7ec8d2d292f48d37e