Add `--bail-on-error` to EqWAlizer commands
Summary:
# Context

When using `elp eqwalize-all` or other eqWAlizer commands in CI, it is desirable to exit with a non-zero status code to signal the fact to users. This point has being brought up a couple of times by the Erlang community. See, for example:

https://github.com/WhatsApp/erlang-language-platform/issues/20

# Solution

Add a `--bail-on-error` flag which makes the eqwalizer commands fail in case a diagnostic is found. Notice that we may want this to be the default behaviour in the future, but for the time being put this behind a flag, to ensure backward compatibility.

Fixes #20.

Reviewed By: VLanvin

Differential Revision: D66768054

fbshipit-source-id: eb4e320fbb474ce341c263bb1f39263b7ac7b89e