Back out "Check for at least OTP 27 when running elp CLI commands that need it"
Summary:
This check was originally introduced to provide better output in the case one would run with OTP 26 an Erlang Service compiled with OTP 27. But this has the consequence of enforcing OTP 27 for the Erlang Service, which should not be necessary and it's breaking CI on GitHub.

Original commit changeset: bd58404544d7

Original Phabricator Diff: D66580167

Reviewed By: alanz

Differential Revision: D66954121

fbshipit-source-id: 5fc3901af567515fe7a864a437bd22e30cded6b3