Favour with_fixture over with_position
Summary:
The `with_position` method was originally a convenience method to quickly access the fixture position.
Over time, it was extended to also return the `diagnostics_enabled` field. But this was rarely used.
Now that we want to expose some other field from the fixture, it feels more natural to just return the fixture itself and explictly access the fields of interest.

Reviewed By: michalmuskala, alanz

Differential Revision: D67145648

fbshipit-source-id: 38d3fb474d575ecdbca78fb05fd641e83b08baff