Add diagnostic tags for unused and deprecated diagnostics
Summary:
LSP version 3.15 brought in the ability to add a tag to a diagnostic, which allows the client to render the item accordingly.

Supported ones are `Unnecessary` and `Deprecated`.

Add this to our diagnostics, and map all the `unused_xxx` diagnostics to `Unnecessary`.

This matches what rust analyzer does, and results in greyed out items for unused, and strikethrough for dperecated.

Reviewed By: VLanvin

Differential Revision: D66965354

fbshipit-source-id: 36fd481436c0e8f7a35a41c44ec2090aa413f22a