Remove Eqwalizer CommandProxy
Summary:
We have been using `CommandProxy` for the eqwalizer `Command` to keep a reference to the Eqwalizer executable on disk, to ensure its drop handler is not invoked while the exe is in use.

But we now have a lazy static reference to the temporary file for the command, which is initialized when ELP starts up and will continue for as long as ELP is running, so we do not need to manage the drop handler for inidividual commands, as there will be a long-lived reference in the lazy static value.

Reviewed By: michalmuskala

Differential Revision: D67193939

fbshipit-source-id: 9167e1a31abc5e02df5213dcf4d63c8765725e7e