update eqwalizer exe if missing
Summary:
# Problem

The eqwalizer executable is bundled inside the ELP exe, and a copy is written to disk in a temporary on startup, and used for the life of the ELP session.  We have been seeing that sometimes this exe disappears, and we assume that it is a long-running ELP LSP server, and the operating system tmp file cleaner deletes it. It does not matter how it disappears, we are seeing that it does.

# Solution

We now thave a single, static, mutable reference to the eqwalizer executable, we add a check prior to attempting to run it that it still exists, and if not we create a fresh version.

Reviewed By: michalmuskala

Differential Revision: D67194324

fbshipit-source-id: c137a51f75ed711e17015ee3033faf2e5d7d27f8