Split struct Eqwalizer so the command is in an inner structure
Summary:
The `Eqwalizer` structure contains information about the exe to run, as well as some config data (the mode to use for Eqwalizer).

Split out the information needed for the executable to be in its own structure, so it can be managed as a unit without worrying about the configuration information.

This will simplify later diffs in the stack which refresh the exe if it vanishes from the file system

Reviewed By: michalmuskala

Differential Revision: D67193907

fbshipit-source-id: 073c210f7e07a112379faf57a54e6c7e69fc924e