erlang_service: use linter abstract forms for ct_info
Summary:
We have switched to using the erlang service linter as the source of truth for OTP-originated diagnostics, and for the abstract forms used in Eqwalizer.  But we have still been using Erlang `compile:file` for evaluating the erlang code for common test lenses.

This diff switches this to instead use the `module_ast` generated abstract forms as used by eqwalizer.

This is the last "OTP" Erlang compilation that does not use ELP for include file resolution.

Reviewed By: michalmuskala

Differential Revision: D67284616

fbshipit-source-id: 58cd99d68600374dd23e4fb4bcebab73eff2566a