Quote module/function/record names in completion, signature help (#71)
Summary:
This improves the display of signature help and fixes label (and edit) values for completions of functions or record names that need quoting, for example `#'basic.publish'{}` from the Rabbit codebase or calling a function on an Elixir module from Erlang.

Also included is a refactor of `base_db::to_quoted_string` to return `Cow<str>` instead, avoiding allocation in the common case that an identifier doesn't need quoting. This saves a few unnecessary allocations.

Pull Request resolved: https://github.com/WhatsApp/erlang-language-platform/pull/71

Reviewed By: michalmuskala

Differential Revision: D67380119

Pulled By: robertoaloi

fbshipit-source-id: 9ecea6b5bd0fff0f127a3a78d25b63f07d8274cf