Make decoding erlang service diagnostics more robust
Summary:
We see occasional panics when attempting to create a `TextRange` for a diagnostic returned from the erlang service where the assert fails, as the start position is not before the end.

Make this safe, by bringing in a function to reverse the order if this occurs when creating the `TextRange`.

e.g. P1699614757

Reviewed By: TD5

Differential Revision: D67392204

fbshipit-source-id: f2f6529c6905c392695ade669539c26f6b27696e