rename: improve check for remote function references
Summary:
Our safety check is currently too strict when checking the rename of a function call in a remote module, in that it checks for a clash whenever there is a call but the function name is already in scope.

But if the call includes the module, then these are in fact safe.

Update the check for this case.

Reviewed By: robertoaloi

Differential Revision: D67450087

fbshipit-source-id: 058bab20d755227a89f610d08c00dc1cec1607bd