Add missing refinement of dynamic/0 into dynamic/1
Summary:
Leverage `dynamic/1` to refine `dynamic/0` in occurrence typing. This effectively solves the issue where the types of all parameters are not refined if one of them is `dynamic()`.

The introduction of new `dynamic/1`'s causes an issue with an existing optimization in `subtype` where we checked for the syntactic presence of a type alias before unfolding it (needed for very large aliases). Fix it by adding `BoundedDynamicType` to this check.

Reviewed By: ilya-klyuchnikov

Differential Revision: D68207950

fbshipit-source-id: f9704adb46cc804297df0c6c19c794721af0a7dc