Add support for rebar3 runnables
Summary:
We currently provide Common Test code lenses only for Buck2 projects. We want to extend this mechanism to rebar3 projects, so that OSS users can eventually use the `edb` debugger via the IDE.

Here we generalize the Runnable so that it can also support rebar3 and create a runnable to execute a test case via `rebar3 ct`. Client support will be added later in the stack.

We do the change in a backward compatible change, thanks to the [`untagged`](https://serde.rs/enum-representations.html#untagged) serde option. This means there is no explicit tag identifying which variant the data contains. Serde will try to match the data against each variant in order and the first one that deserializes successfully is the one returned.

Reviewed By: alanz

Differential Revision: D68106799

fbshipit-source-id: 1e241e6e98214aa321b168cfd7b7c5605828555f