2/n: ssr: Introduce SsrSource as interned source for Ssr definitions
Summary:
This allows us to add SSR definitions in assist code, and reference them in salsa without having to mutate the database.

Most of these will originate in code or loaded from a config file and so will be of a static size.

The occasional ad-hoc one via a command or comment should not cause a memory blow up.

Reviewed By: michalmuskala

Differential Revision: D68215148

fbshipit-source-id: b72b881722d08691ef3294defa0aada88cfe6d2f