Allow configuring multiple apps/deps patterns via .elp.toml
Summary:
When configuring a custom repository it may be convenient to specify multiple `apps` and `deps` pattern in the `build_info` section of the `.elp.toml` file.

We currently only support a single string.

This change adds support for an array of strings as an alternative. For backward compatibility, I'm keeping the old format. We should consider deprecating that, eventually.

Reviewed By: alanz

Differential Revision: D68489526

fbshipit-source-id: 5c12eb0e4c3f29d33b79b1123732ee33ddc09251