BE: Use only backticks in config markdown help text
Summary:
We use the output of `elp config` to update the emacs lsp-mode configuration options for the Erlang client.

This is done via an elisp script. But that script [generates warnings](https://github.com/emacs-lsp/lsp-mode/pull/4637#pullrequestreview-2530298803) if we use single quotes on items in our markdown help text.

So remove the only instance of this in our help text, harmonising to use backquotes everywhere in the process.

Reviewed By: jcpetruzza

Differential Revision: D68491031

fbshipit-source-id: 3ae3d84528e737903ef0c50fab3a275bba584e0d