20/n: ssr: match Expr::Case
Summary:
As title.

We match the expression that is the subject of the case, then a flattening of all the given cases. At this stage we cannot pick out a single case clause and ignore the rest.

For the case clauses (`CRClause`), we explictly insert markers to delimit the guards and expressions, to make sure we do not get accidental matches in the wrong place.

Reviewed By: robertoaloi

Differential Revision: D68274787

fbshipit-source-id: 0aead78fcb94b4777e83d34b8be24491828458ec