9/n: ssr: match on record field names too
Summary:
Up to now we have matched the children of a HIR AST item as a linear, ordered list.

For matching items such as maps or records, we need to be able to specify that things must match, but in any order.

We do this by introducing `PaternIterator::Map`. This captures a fixed prefix, for the record or field name, and a hash map of items that need to be matched as key value pairs.

The check for a match then considers them in any order.

This does not yet allow us to specify a subset match, ie look for a map reference with only some keys specified.

Reviewed By: robertoaloi

Differential Revision: D68264859

fbshipit-source-id: 5cc804de0eaac11bb197c955fcf2fe169c7ffe20