5/n: ssr: Add first match
Summary:
# Overview

Matching proceeds by comparing the HIR AST of the pattern with that of the code.
We are potentially comparing disparate things though, as we descend the tree
- a normal HIR Pat or Exp
- an atom
- a binary op
- a unary op
- a map op

We keep track of what precisely we are looking at using a `SubId`, and flatten the comparisons by having a `variant_str` method which returns a static string for each, which we compare.

This has the nice side-property of making debugging of the match process simpler, as we can print the strings and have human-readable information.  It may be possible to switch this to some unique number process instead, but that is a mechanical transformation if performance is an issue later.

For each HIR node having children, we invoke a `PatternIterator` to produce a matchable representation.  Since the HIR AST removes some structure, we reintroduce it in this as needed.  The matchable representation is initially just a `Vec<SubId>`, but further up the stack we cope with arbitrary order too for fields and maps.

And if the pattern has a placeholder, we just accept whatever is currently at the same position for the code.

The search only matches one pattern at a time, matching in parallel is not a solved problem.

# This Diff

This diff does the initial match infrastructure, for the first simple match, and leaves most of the specific match information as `todo!()` macros.  These will be fleshed out as the stack progresses.

The structure of the code is based very strongly on what is in rust analyzer.

Reviewed By: robertoaloi

Differential Revision: D68218409

fbshipit-source-id: 4b4bd9cb8218b453c373b79fa829d9761a8b1c3c