8/n: ssr: Add match on record
Summary:
Add a match on a record.  This does not yet deal with arbitrary order of the presentation of the record contents.

Note also the technique for matching the binary, where we create a linear list of `SubId`s, concatenating those from each of the sub-components of the binary.

This is a general technique we use, and in some cases choose the order explicitly to fail fast if e.g. an operator does not match by always putting it first.

The key point is that both sides of the match use the same ordering, so the match will work whatever we use.

Reviewed By: robertoaloi

Differential Revision: D68263753

fbshipit-source-id: 2700c0fd4e5e08f13904fcf6828f7ce812bc69b3