4/n: ssr: introduce ide_ssr crate
Summary:
The structural search and replace functionality will live in the ide_ssr crate.

This is based strongly on the [equivalent in rust-analyzer](https://github.com/rust-lang/rust-analyzer/tree/master/crates/ide-ssr), including its description which is purely aspirational at this stage of the stack.

The initial version is just a bare minimum, but does include tests for parsing, lowering and tree printing some SSR definitions.

Reviewed By: jcpetruzza, michalmuskala, robertoaloi

Differential Revision: D68216807

fbshipit-source-id: da1e29191088f897d41a04344b7f0debd1771783