41/n: ssr: BE: improve PatternIterator::new_any_expr
Summary:
This method does case analysis on an AnyExpr, and constructs a `PatternIterator` for each variant.  We have two kinds of `PatternIterator`, one for matching linear lists of variant children, and one matching children where some may come in any order, for matching the contents of maps or records.

This has been done by returning an `either` type from the match, and converting it to the specific iterator at the end.

This diff instead introduces auxiliary methods to construct the `PatternIterator` variants, and uses these directly.

Reviewed By: robertoaloi

Differential Revision: D68555752

fbshipit-source-id: a67fbb0821870f3095a31cfbae74a4fecd3a5d86