37/n: ssr: Initial processing of when guards
Summary:
Add initial processing of the `when` clause in an SSR definition.

These are used to specify conditions on the placeholder matches.  The initial check is for `==` and `/=` to a literal.  This can be useful when matching a specific atom, as the result includes the location and identity of any matched placeholders, so having this simplifies use of the match.

e.g.

```
ssr: {_Foo, _X} when _Foo == foo.
```

vs
```
ssr: {foo, _X}.
```

They will both match the same tuples, but we get the range of `_Foo` "for free" in the match result.

Reviewed By: robertoaloi

Differential Revision: D68491445

fbshipit-source-id: fb0326cdbb5570f516e43c0625f25800ee96d853