44/n: ssr: record macro call name in HIR
Summary:
When we lower a macro call in HIR we only store the resolved macro, if it resolves.

An SSR pattern is lowered in isolation, and if it has a macro in it we do not have its name.

So store the name in the HIR `MacroCall` variants, so it can be used for matching.

Reviewed By: robertoaloi

Differential Revision: D68562000

fbshipit-source-id: 51973f1a7b16bd12a9486ccf57bdd0dbaed523ca