42/n: ssr: use fold::Strategy when doing a match
Summary:
When writing a diagnostic or assist, we sometimes want to expand macros, sometimes not. And sometimes we want to be aware of parens, and sometimes not.

This is normally set via a `fold::Strategy` parameter passed to the fold doing the search.

This diff introduces it for the ssr matching too. It is used to control the underlying fold, but also to control how the search process works.  This means search patterns can be searched where parens matter or not, as needed.

Reviewed By: robertoaloi

Differential Revision: D68555952

fbshipit-source-id: ff4af56dc5059e1f9584b7a7cd7b5cc708c70577